Completed Multi-Select Intake, Province & Level
Created rearrange_dictionary to my priority as the user.

Worked for priority function to make it work as it should in the finest way possible.

Worked on the preparing a proper data to provide the CRS for Better and Enhanced Recommendation of courses

Also Changes college_df to new_college_df for less confusion and better readability.